DEV-2387 Separate file name validations on windows vs. unix (#306)
* DEV-2387 Separate file name validations on windows vs. unix

Windows disallows certain characters in file names that Unix does not.
This is required because the AgentServer uses the local filesystem for
file storage in Studio. However, we should not over-enforce this limit
as it creates a situation where ACE/SPAR/etc then cannot store files
with names that the OS does allow.

This change separates the `_check_that_name_is_valid_in_filesystem` function
into unix and windows equivalent functions, delegating to the appropriate
one depending on platform.

* update docs

* formatting

* Switch from platform.system() to sys.platform